+++ lib/fetch/Fetch.d.ts
@@ -32,6 +32,14 @@ export default class Fetch {
      */
     audio(key: string): (HTMLElement | undefined);
     /**
+     * Returns an audio buffer from the cache.
+     *
+     * @param {string} key The key of the audio buffer to return.
+     *
+     * @returns {AudioBuffer} Returns the audio buffer.
+     */
+    audioBuffer(key: string): (AudioBuffer | undefined);
+    /**
      * Returns a video asset from the cache.
      *
      * @param {string} key The key of the video asset to return.
@@ -63,4 +71,12 @@ export default class Fetch {
      * @returns {HTMLElement} Returns the json asset.
      */
     json(key: string): (HTMLElement | undefined);
+    /**
+     * Returns an array buffer from the cache.
+     *
+     * @param {string} key The key of the array buffer to return.
+     *
+     * @returns {ArrayBuffer} Returns the array buffer.
+     */
+    arrayBuffer(key: string): (ArrayBuffer | undefined);
 }