+++ lib/options/Options.d.ts
@@ -1,0 +1,25 @@
+/**
+ * Defines the options for an instance of MuskOx and their default values.
+ */
+export default class Options {
+    /**
+     * A cross-origin property to set for all assets that use cross-origin.
+     *
+     * @property {string}
+     *
+     * @default ''
+     */
+    crossOrigin: string;
+    /**
+     * A reference to an existing AudioContext to use if creating web audio assets.
+     *
+     * If one is not assigned then a new instance of an AudioContext will be used.
+     *
+     * @property {AudioContext}
+     */
+    audioContext: (AudioContext | null);
+    /**
+     * @param {Object} options The options passed to MuskOx on initialization.
+     */
+    constructor(options: Object);
+}