+++ muskox.js
@@ -87,6 +87,19 @@ function () {
       return this._cache.get('audio', key);
     }
     /**
+     * Returns an audio buffer from the cache.
+     * 
+     * @param {string} key The key of the audio buffer to return.
+     * 
+     * @returns {AudioBuffer} Returns the audio buffer.
+     */
+
+  }, {
+    key: "audioBuffer",
+    value: function audioBuffer(key) {
+      return this._cache.get('audioBuffer', key);
+    }
+    /**
      * Returns a video asset from the cache.
      * 
      * @param {string} key The key of the video asset to return.
@@ -138,6 +151,20 @@ function () {
     value: function json(key) {
       return this._cache.get('json', key);
     }
+    /**
+     * Returns an array buffer from the cache.
+     * 
+     * @param {string} key The key of the array buffer to return.
+     * 
+     * @returns {ArrayBuffer} Returns the array buffer.
+     */
+
+  }, {
+    key: "arrayBuffer",
+    value: function arrayBuffer(key) {
+      // @ts-ignore
+      return this._cache.get('arrayBuffer', key);
+    }
   }]);
 
   return Fetch;
@@ -155,7 +182,9 @@ function () {
       video: {},
       text: {},
       binary: {},
-      json: {}
+      json: {},
+      arrayBuffer: {},
+      audioBuffer: {}
     });
   }
 
@@ -277,6 +306,37 @@ function getPlayableMedia(type, srcs) {
 
   return '';
 }
+
+var Options =
+/**
+ * A cross-origin property to set for all assets that use cross-origin.
+ * 
+ * @property {string}
+ * 
+ * @default ''
+ */
+
+/**
+ * A reference to an existing AudioContext to use if creating web audio assets.
+ * 
+ * If one is not assigned then a new instance of an AudioContext will be used.
+ * 
+ * @property {AudioContext}
+ */
+
+/**
+ * @param {Object} options The options passed to MuskOx on initialization.
+ */
+function Options(options) {
+  _classCallCheck(this, Options);
+
+  _defineProperty(this, "crossOrigin", '');
+
+  _defineProperty(this, "audioContext", null);
+
+  Object.assign(this, options);
+  if (!this.audioContext) this.audioContext = new AudioContext();
+};
 
 function _classCallCheck$1(instance, Constructor) {
   if (!(instance instanceof Constructor)) {
@@ -728,6 +788,14 @@ var MuskOx =
 /*#__PURE__*/
 function () {
   /**
+   * A reference to the options for this instance of MuskOx.
+   * 
+   * @private
+   * 
+   * @property {Options}
+   */
+
+  /**
    * A reference to the cache used to store assets.
     * 
     * @private
@@ -765,14 +833,6 @@ function () {
    * @property {number}
     * 
     * @private
-   */
-
-  /**
-   * The crossOrigin option passed to MuskOx on initialization.
-   * 
-   * @private
-   * 
-   * @property {string}
    */
 
   /**
@@ -822,12 +882,16 @@ function () {
    */
 
   /**
-   * @param {string} crossOrigin The crossOrigin option passed to MuskOx on initialization.
+   * @param {Object} [options]
+    * @param {string} [options.crossOrigin=''] A cross-origin property to set for all assets that use cross-origin.
+    * @param {AudioContext} [options.audioContext=new AudioContext()] A reference to an existing AudioContext to use if creating web audio assets. If one is not assigned then a new instance of an AudioContext will be used.
    */
   function MuskOx() {
-    var crossOrigin = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : '';
+    var options = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {};
 
     _classCallCheck(this, MuskOx);
+
+    _defineProperty(this, "_options", void 0);
 
     _defineProperty(this, "_cache", new Cache());
 
@@ -839,8 +903,6 @@ function () {
 
     _defineProperty(this, "_assetsToLoad", 0);
 
-    _defineProperty(this, "_crossOrigin", void 0);
-
     _defineProperty(this, "_progress", 0);
 
     _defineProperty(this, "_onProgress", new Hypergiant());
@@ -851,7 +913,7 @@ function () {
 
     _defineProperty(this, "_onComplete", new Hypergiant());
 
-    this._crossOrigin = crossOrigin;
+    this._options = new Options(options);
   }
   /**
    * Returns the cache module.
@@ -891,6 +953,8 @@ function () {
             case 'text':
             case 'binary':
             case 'json':
+            case 'arrayBuffer':
+            case 'audioBuffer':
               this._loadXHR(asset);
 
               break;
@@ -929,7 +993,7 @@ function () {
     /**
      * Adds an audio asset to the load queue.
      * 
-     * Muliple `src` paths can be provided in case one or more are not supported by the user's browser.
+     * Multiple `src` paths can be provided in case one or more are not supported by the user's browser.
      * 
      * @param {string} key A unique key to reference this audio asset by.
      * @param {string|Array<string>} src A path to the audio asset or an array of paths to an audio asset and its fallbacks.
@@ -942,6 +1006,21 @@ function () {
       var replace = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : false;
 
       this._addToQueue('audio', key, srcs, replace);
+    }
+    /**
+     * Adds an AudioBuffer to the load queue.
+     * 
+     * @param {string} key A unique key to reference this audio buffer by.
+    * @param {string} src A path to the audio asset.
+    * @param {boolean} [replace=false] Indicates whether an audio buffer with the same key should be replaced in the cache or not.
+     */
+
+  }, {
+    key: "audioBuffer",
+    value: function audioBuffer(key, src) {
+      var replace = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : false;
+
+      this._addToQueue('audioBuffer', key, src, replace);
     }
     /**
      * Adds a video asset to the load queue.
@@ -1006,6 +1085,23 @@ function () {
       this._addToQueue('json', key, src, replace);
     }
     /**
+     * Loads an asset as an array buffer.
+     * 
+     * This can be useful for loading an audio asset to use with web audio.
+     * 
+     * @param {string} key A unique key to reference this array buffer asset by.
+     * @param {string} src The path to the asset.
+     * @param {boolean} [replace=false] Indicates whether an array buffer asset with the same key should be replaced in the cache or not.
+     */
+
+  }, {
+    key: "arrayBuffer",
+    value: function arrayBuffer(key, src) {
+      var replace = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : false;
+
+      this._addToQueue('arrayBuffer', key, src, replace);
+    }
+    /**
      * Takes the supplied asset, creates an asset instance out of it, and
      * adds it to the load queue.
       * 
@@ -1052,7 +1148,7 @@ function () {
         _this._handleAssetError(asset, err);
       }, false);
       asset.data.src = asset.src.toString();
-      if (this._crossOrigin) asset.data.crossOrigin = this._crossOrigin;
+      if (this._options.crossOrigin) asset.data.crossOrigin = this._options.crossOrigin;
     }
     /**
      * Load assets that can be loaded through the use of the `canPlayThrough` event
@@ -1092,30 +1188,51 @@ function () {
       var _this3 = this;
 
       asset.data = new XMLHttpRequest();
+      var loaded = new Hypergiant();
+      loaded.add(function () {
+        return _this3._cacheAsset(asset);
+      });
       asset.data.addEventListener('readystatechange', function () {
         if (asset.data.readyState == 4 && asset.data.status == 200) {
           switch (asset.type) {
             case 'text':
               asset.data = asset.data.responseText;
+              loaded.dispatch();
               break;
 
             case 'binary':
               var arrayBuffer = asset.data.response;
               if (arrayBuffer) asset.data = new Uint8Array(arrayBuffer);
+              loaded.dispatch();
               break;
 
             case 'json':
               asset.data = JSON.parse(asset.data.responseText);
+              loaded.dispatch();
               break;
-          }
 
-          _this3._cacheAsset(asset);
+            case 'arrayBuffer':
+              asset.data = asset.data.response;
+              loaded.dispatch();
+              break;
+
+            case 'audioBuffer':
+              var response = asset.data.response;
+
+              _this3._options.audioContext.decodeAudioData(response, function (buffer) {
+                asset.data = buffer;
+                loaded.dispatch();
+              });
+
+              break;
+          } //this._cacheAsset(asset);
+
         }
       }, false);
       asset.data.addEventListener('error', function (err) {
         _this3._handleAssetError(asset, err);
       }, false);
-      if (asset.type == 'binary') asset.data.responseType = 'arraybuffer';
+      if (asset.type == 'binary' || asset.type === 'arrayBuffer' || asset.type === 'audioBuffer') asset.data.responseType = 'arraybuffer';
       asset.data.open('GET', asset.src);
       asset.data.send();
     }