+++ src/cache/Cache.ts
@@ -15,7 +15,9 @@ export default class Cache {
     video: {},
     text: {},
     binary: {},
-    json: {}
+		json: {},
+		arrayBuffer: {},
+		audioBuffer: {},
   };
 
 	/**
@@ -26,7 +28,7 @@ export default class Cache {
 	 * 
 	 * @returns {HTMLElement|undefined} Returns the asset or undefined if it doesn't exist.
 	 */
-  get(type: string, key: string): (HTMLElement | undefined) {
+  get(type: string, key: string): any {
     if (this._exists(type, key)) return this.assets[type][key];
   }
 
@@ -40,10 +42,10 @@ export default class Cache {
 	 * 
 	 * @returns {boolean} Returns true if the asset was successfully added to the cache or false if the asset already exists and `replace` is set to false.
 	 */
-  set(type: string, key: string, asset: HTMLElement, replace: boolean = false): boolean {
+  set(type: string, key: string, asset: any, replace: boolean = false): boolean {
     if (this._exists(type, key) && !replace) return false;
 
-    this.assets[type][key] = asset;
+		this.assets[type][key] = asset;
 
     return true;
   }